+++ lib/nadia.ex
@@ -3130,6 +3130,158 @@ defmodule Nadia do
   end
 
   @doc """
+  Use this method to reply to shipping queries.
+  Returns `:ok` on success.
+
+  Args:
+  * `shipping_query_id` - Unique identifier for the query to be answered
+  * `ok` - Pass true if delivery to the specified address is possible
+  * `options` - orddict of options
+
+  Options:
+  * `:shipping_options` - JSON-serializable list of shipping options
+  * `:error_message` - Error message to display when `ok` is false
+  """
+  @spec answer_shipping_query(binary, boolean) :: :ok | {:error, Error.t()}
+  @spec answer_shipping_query(binary, boolean, [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  @spec answer_shipping_query(Client.t(), binary, boolean) :: :ok | {:error, Error.t()}
+  @spec answer_shipping_query(Client.t(), binary, boolean, [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  def answer_shipping_query(shipping_query_id, ok) do
+    answer_shipping_query(shipping_query_id, ok, [])
+  end
+
+  def answer_shipping_query(%Client{} = client, shipping_query_id, ok) do
+    answer_shipping_query(client, shipping_query_id, ok, [])
+  end
+
+  def answer_shipping_query(shipping_query_id, ok, options) do
+    api_request(
+      "answerShippingQuery",
+      request_options(
+        [shipping_query_id: shipping_query_id, ok: ok],
+        encode_json_option(options, :shipping_options)
+      )
+    )
+  end
+
+  def answer_shipping_query(%Client{} = client, shipping_query_id, ok, options) do
+    api_request(
+      client,
+      "answerShippingQuery",
+      request_options(
+        [shipping_query_id: shipping_query_id, ok: ok],
+        encode_json_option(options, :shipping_options)
+      )
+    )
+  end
+
+  @doc """
+  Use this method to respond to pre-checkout queries.
+  Returns `:ok` on success.
+
+  Args:
+  * `pre_checkout_query_id` - Unique identifier for the query to be answered
+  * `ok` - Pass true if the bot is ready to proceed with the order
+  * `options` - orddict of options
+
+  Options:
+  * `:error_message` - Error message to display when `ok` is false
+  """
+  @spec answer_pre_checkout_query(binary, boolean) :: :ok | {:error, Error.t()}
+  @spec answer_pre_checkout_query(binary, boolean, [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  @spec answer_pre_checkout_query(Client.t(), binary, boolean) :: :ok | {:error, Error.t()}
+  @spec answer_pre_checkout_query(Client.t(), binary, boolean, [{atom, any}] | map) ::
+          :ok | {:error, Error.t()}
+  def answer_pre_checkout_query(pre_checkout_query_id, ok) do
+    answer_pre_checkout_query(pre_checkout_query_id, ok, [])
+  end
+
+  def answer_pre_checkout_query(%Client{} = client, pre_checkout_query_id, ok) do
+    answer_pre_checkout_query(client, pre_checkout_query_id, ok, [])
+  end
+
+  def answer_pre_checkout_query(pre_checkout_query_id, ok, options) do
+    api_request(
+      "answerPreCheckoutQuery",
+      request_options([pre_checkout_query_id: pre_checkout_query_id, ok: ok], options)
+    )
+  end
+
+  def answer_pre_checkout_query(%Client{} = client, pre_checkout_query_id, ok, options) do
+    api_request(
+      client,
+      "answerPreCheckoutQuery",
+      request_options([pre_checkout_query_id: pre_checkout_query_id, ok: ok], options)
+    )
+  end
+
+  @doc """
+  Use this method to refund a successful Telegram Stars payment.
+  Returns `:ok` on success.
+
+  Args:
+  * `user_id` - Identifier of the user whose payment will be refunded
+  * `telegram_payment_charge_id` - Telegram payment identifier
+  """
+  @spec refund_star_payment(integer, binary) :: :ok | {:error, Error.t()}
+  @spec refund_star_payment(Client.t(), integer, binary) :: :ok | {:error, Error.t()}
+  def refund_star_payment(user_id, telegram_payment_charge_id) do
+    api_request(
+      "refundStarPayment",
+      user_id: user_id,
+      telegram_payment_charge_id: telegram_payment_charge_id
+    )
+  end
+
+  def refund_star_payment(%Client{} = client, user_id, telegram_payment_charge_id) do
+    api_request(
+      client,
+      "refundStarPayment",
+      user_id: user_id,
+      telegram_payment_charge_id: telegram_payment_charge_id
+    )
+  end
+
+  @doc """
+  Use this method to cancel or re-enable extension of a Telegram Stars subscription.
+  Returns `:ok` on success.
+
+  Args:
+  * `user_id` - Identifier of the user whose subscription will be edited
+  * `telegram_payment_charge_id` - Telegram payment identifier for the subscription
+  * `is_canceled` - Pass true to cancel extension, or false to re-enable it
+  """
+  @spec edit_user_star_subscription(integer, binary, boolean) :: :ok | {:error, Error.t()}
+  @spec edit_user_star_subscription(Client.t(), integer, binary, boolean) ::
+          :ok | {:error, Error.t()}
+  def edit_user_star_subscription(user_id, telegram_payment_charge_id, is_canceled) do
+    api_request(
+      "editUserStarSubscription",
+      user_id: user_id,
+      telegram_payment_charge_id: telegram_payment_charge_id,
+      is_canceled: is_canceled
+    )
+  end
+
+  def edit_user_star_subscription(
+        %Client{} = client,
+        user_id,
+        telegram_payment_charge_id,
+        is_canceled
+      ) do
+    api_request(
+      client,
+      "editUserStarSubscription",
+      user_id: user_id,
+      telegram_payment_charge_id: telegram_payment_charge_id,
+      is_canceled: is_canceled
+    )
+  end
+
+  @doc """
   Use this method to get the token of a managed bot.
   Returns the token as a string.